remove direct dep on cerd from aziot
